Fix material IPR updates, added markTransformsChildrenAsDirty to the rendercallbacks
